fix: Address review comments and add upgrades
