+++ py/torch_tensorrt/dynamo/lowering/__init__.py
@@ -2,5 +2,5 @@ from ._decompositions import get_decompositions  # noqa: F401
 from ._fusers import *  # noqa: F401
 from ._pre_aot_lowering import SUBSTITUTION_REGISTRY  # noqa: F401
 from ._pre_aot_lowering import register_substitution  # noqa: F401
-from .passes import add_lowering_pass, apply_lowering_passes
+from .passes import apply_lowering_passes
 from .substitutions import *  # noqa: F401
